refactor themeRepository codes
