+++ electron-app/src/renderer/repositories/theme/index.ts
@@ -1,0 +1,4 @@
+export * from './browserThemeImpl'
+export * from './electronThemeImpl'
+export * from './registry'
+export * from './repository'
