+++ electron-app/src/renderer/repositories/theme/repository.ts
@@ -1,0 +1,7 @@
+import { ThemePreferredType, ThemeType } from '../../types/theme/type'
+
+export interface Repository {
+  setPreferredTheme(theme: ThemePreferredType): void
+  getPreferredTheme(): ThemePreferredType
+  getTheme(): ThemeType
+}
