+++ electron-app/src/renderer/types/theme/type.ts
@@ -1,0 +1,3 @@
+export type ThemeType = 'dark' | 'light'
+export type ThemePreferredType = ThemeType | 'system'
+export type ThemeStorageType = 'browser' | 'electron'
